MWPW-165786-split personalization to separate dme from common things, dme constants
