--- edsdme/blocks/search-full/cards.js
@@ -1,266 +1,0 @@
-const cardsData = {
-  cards: [
-    {
-      id: '1',
-      styles: {
-        backgroundImage: 'https://main--dme-partners--adobecom.hlx.live/default-meta-image.png?width=1200&format=pjpg&optimize=medium',
-        backgroundAltText: 'backgroundAltText',
-      },
-      contentArea: {
-        detailText: '',
-        title: 'Captivate 2019, Channel Copy Existing Users (German)',
-        url: 'https://partnerdownload.adobe.com/p/captivate_2019_channel_copy_existing_users_d.pdf',
-        description: 'Captivate 2019, Channel Copy Existing Users',
-        size: '1 MB',
-        type: 'pdf',
-        contentType: 'page',
-      },
-      cardDate: '2024-06-09T12:35:03.000Z',
-      modifiedDate: '2024-06-09T12:35:03.000Z',
-      arbitrary: [
-        {
-          key: 'topic',
-          value: 'analyst-reports',
-        },
-        {
-          key: 'topic',
-          value: 'marketing',
-        },
-      ],
-      lang: 'en',
-      country: 'us',
-      origin: 'dme-partners',
-    },
-    {
-      id: '2',
-      styles: {
-        backgroundImage: '',
-        backgroundAltText: '',
-      },
-      contentArea: {
-        detailText: '',
-        title: 'Premiereï¿½Elements 2018 Boxshot, 3D (UE)',
-        url: 'https://partnerdownload.adobe.com/p/MHAsset-216593.zip',
-        description: 'Captivate 2019, Channel Copy Existing Users',
-        size: '88 MB',
-        type: 'zip',
-        contentType: 'asset',
-      },
-      cardDate: '2024-07-09T12:35:03.000Z',
-      modifiedDate: '2024-07-09T12:35:03.000Z',
-      arbitrary: [
-        {
-          key: 'topic',
-          value: 'analyst-reports',
-        },
-      ],
-      lang: 'en',
-      country: 'us',
-      origin: 'dme-partners',
-    },
-    {
-      id: '3',
-      styles: {
-        backgroundImage: 'https://main--dme-partners--adobecom.hlx.live/default-meta-image.png?width=1200&format=pjpg&optimize=medium',
-        backgroundAltText: 'backgroundAltText',
-      },
-      contentArea: {
-        detailText: '',
-        title: 'Adobe Partner Connection Online Enrollment Checklist Certified/Gold - China (CN)',
-        url: 'https://partnerdownload.adobe.com/p/Online_Enrollment_Cert_Gold_v1.0_CN.pdf',
-        description: 'Checklist covering the online enrollment process for the Adobe Partner Connection Certified and Gold levels.',
-        size: '493 KB',
-        type: 'm4v',
-        contentType: 'page',
-      },
-      cardDate: '2024-05-09T12:35:03.000Z',
-      modifiedDate: '2024-05-09T12:35:03.000Z',
-      arbitrary: [
-        {
-          key: 'topic',
-          value: 'amplify',
-        },
-      ],
-      lang: 'en',
-      country: 'us',
-      origin: 'dme-partners',
-    },
-    {
-      id: '4',
-      styles: {
-        backgroundImage: 'https://main--dme-partners--adobecom.hlx.live/default-meta-image.png?width=1200&format=pjpg&optimize=medium',
-        backgroundAltText: 'backgroundAltText',
-      },
-      contentArea: {
-        detailText: '',
-        title: 'Adobe Partner Connection Online Enrollment Checklist Registered - China (CN)',
-        url: 'https://partnerdownload.adobe.com/p/Online_Enrollment_Registered_v1.0_CN.pdf',
-        description: 'Checklist covering the online enrollment process for the Adobe Partner Connection registered level.',
-        size: '452 KB',
-        type: 'word',
-        contentType: 'page',
-      },
-      cardDate: '2024-04-09T12:35:03.000Z',
-      modifiedDate: '2024-04-09T12:35:03.000Z',
-      arbitrary: [
-        {
-          key: 'topic',
-          value: 'amplify',
-        },
-      ],
-      lang: 'en',
-      country: 'us',
-      origin: 'dme-partners',
-    },
-    {
-      id: '5',
-      styles: {
-        backgroundImage: 'https://main--dme-partners--adobecom.hlx.live/default-meta-image.png?width=1200&format=pjpg&optimize=medium',
-        backgroundAltText: 'backgroundAltText',
-      },
-      contentArea: {
-        detailText: '',
-        title: 'Restricting Access and Permissions Arabic AR',
-        url: 'https://partnerdownload.adobe.com/p/restricting_access_and_permissions--ar.pdf',
-        description: 'Restricting Access and Permissions - Arabic',
-        size: '80 KB',
-        type: 'xlsx',
-        contentType: 'page',
-      },
-      cardDate: '2024-03-09T12:35:03.000Z',
-      modifiedDate: '2024-03-09T12:35:03.000Z',
-      arbitrary: [
-        {
-          key: 'topic',
-          value: 'amplify',
-        },
-      ],
-      lang: 'en',
-      country: 'us',
-      origin: 'dme-partners',
-    },
-    {
-      id: '6',
-      styles: {
-        backgroundImage: 'https://main--dme-partners--adobecom.hlx.live/default-meta-image.png?width=1200&format=pjpg&optimize=medium',
-        backgroundAltText: 'backgroundAltText',
-      },
-      contentArea: {
-        detailText: '',
-        title: 'Adobe Partner Connection Reseller Program Datasheet - Latin America (ES)',
-        url: 'https://partnerdownload.adobe.com/p/LAM_Reseller_Program_Datasheet_ES.pdf',
-        description: 'Datasheet covering the requirements and benefits of the Reseller Program.',
-        size: '120 KB',
-        type: 'ppt',
-        contentType: 'page',
-      },
-      cardDate: '2024-02-09T12:35:03.000Z',
-      modifiedDate: '2024-02-09T12:35:03.000Z',
-      arbitrary: [
-        {
-          key: 'topic',
-          value: 'analyst-reports',
-        },
-      ],
-      lang: 'en',
-      country: 'us',
-      origin: 'dme-partners',
-    },
-    {
-      id: '7',
-      styles: {
-        backgroundImage: '',
-        backgroundAltText: '',
-      },
-      contentArea: {
-        detailText: '',
-        title: 'Photoshop Elements and Premiereï¿½Elements 2018 Boxshot, 3D (UE)',
-        url: 'https://partnerdownload.adobe.com/p/MHAsset-216587.zip',
-        description: 'Marketing materials for Photoshop Elements & Premiere Elements 2018 product launch.',
-        size: '84 MB',
-        type: 'rec',
-        contentType: 'asset',
-      },
-      cardDate: '2024-01-09T12:35:03.000Z',
-      modifiedDate: '2024-01-09T12:35:03.000Z',
-      arbitrary: [
-        {
-          key: 'topic',
-          value: 'analyst-reports',
-        },
-        {
-          key: 'topic',
-          value: 'marketing',
-        },
-      ],
-      lang: 'en',
-      country: 'us',
-      origin: 'dme-partners',
-    },
-    {
-      id: '8',
-      styles: {
-        backgroundImage: 'https://main--dme-partners--adobecom.hlx.live/default-meta-image.png?width=1200&format=pjpg&optimize=medium',
-        backgroundAltText: 'backgroundAltText',
-      },
-      contentArea: {
-        detailText: '',
-        title: 'Securing Your PDF Portfolio (Arabic)',
-        url: 'https://partnerdownload.adobe.com/p/securing_your_pdf_portfolio--ar.pdf',
-        description: 'Securing Your PDF Portfolio - Arabic',
-        size: '106 KB',
-        type: 'html',
-        contentType: 'page',
-      },
-      cardDate: '2023-07-09T12:35:03.000Z',
-      modifiedDate: '2023-07-09T12:35:03.000Z',
-      arbitrary: [
-        {
-          key: 'topic',
-          value: 'analyst-reports',
-        },
-      ],
-      lang: 'en',
-      country: 'us',
-      origin: 'dme-partners',
-    },
-    {
-      id: '9',
-      styles: {
-        backgroundImage: 'https://main--dme-partners--adobecom.hlx.live/default-meta-image.png?width=1200&format=pjpg&optimize=medium',
-        backgroundAltText: 'backgroundAltText',
-      },
-      contentArea: {
-        detailText: '',
-        title: 'Captivate 2020, Channel Copy Existing Users',
-        url: 'https://partnerdownload.adobe.com/p/captivate_2019_channel_copy_existing_users_d.pdf',
-        description: 'Captivate 2019, Channel Copy Existing Users',
-        size: '1 MB',
-        type: 'img',
-        contentType: 'page',
-      },
-      cardDate: '2024-06-09T12:35:03.000Z',
-      modifiedDate: '2024-06-09T12:35:03.000Z',
-      arbitrary: [
-        {
-          key: 'topic',
-          value: 'analyst-reports',
-        },
-        {
-          key: 'topic',
-          value: 'marketing',
-        },
-      ],
-      lang: 'en',
-      country: 'us',
-      origin: 'dme-partners',
-    },
-  ],
-  filters: [],
-  debug: [
-    { missingIds: [] },
-  ],
-  isHashed: false,
-};
-
-export default cardsData;
