+++ edsdme/blocks/utils/dmeConstants.js
@@ -1,0 +1,1 @@
+export const RT_SEARCH_ACTION_PATH = '/api/v1/web/dx-partners-runtime/search-apc/search-apc?';
