+++ edsdme/scripts/personalizationUtils.js
@@ -1,0 +1,25 @@
+import {COOKIE_OBJECT, PERSONALIZATION_HIDE} from "./personalizationConfig.js";
+import {hasSalesCenterAccess} from "./utils.js";
+
+export  function processPrimaryContact(el) {
+  const isPrimary = COOKIE_OBJECT.primaryContact;
+  el.classList.add(PERSONALIZATION_HIDE);
+  if (!isPrimary) return;
+  const primaryContactWrapper = document.createElement('div');
+  const primaryContact = document.createElement('p');
+  primaryContact.textContent = el.textContent;
+  primaryContactWrapper.classList.add('primary-contact-wrapper');
+  primaryContactWrapper.appendChild(primaryContact);
+  el.replaceWith(primaryContactWrapper);
+}
+
+export function processSalesAccess(el) {
+  const salesAccess = hasSalesCenterAccess();
+  const element = el.parentElement;
+  if (!salesAccess) {
+    element.classList.add(PERSONALIZATION_HIDE);
+    return;
+  }
+  const divider = document.createElement('hr');
+  element.insertBefore(divider, el);
+}
